api/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# Supabase
supabase>=2.3.0

# AI/ML Core
torch>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0

# Image Embeddings (CLIP)
open-clip-torch>=2.24.0

# Object Detection (YOLO)
ultralytics>=8.1.0

# Face Recognition
insightface>=0.7.3
onnxruntime>=1.16.0

# OCR
easyocr>=1.7.0

# Vision-Language Model
transformers>=4.36.0
accelerate>=0.25.0

# Clustering
hdbscan>=0.8.33
scikit-learn>=1.3.0

# Image utilities
imagehash>=4.3.0
opencv-python-headless>=4.8.0

# Task queue (for batch processing)
celery>=5.3.0
redis>=5.0.0

# Utilities
python-multipart>=0.0.6
python-jose>=3.3.0
watchdog>=3.0.0  # For hot-reload in development
